update dark theme for browser
